Defensive programming to prevent integer overflow for large images (#2232)
Sometimes for very large images, either the number of pixels (w * h * d)
or the total data size (npixels * pixelbytes) can overflow int32 math.
Find several places where this may be a problem and use casts to ensure
64 bit multiplication.

I don't claim this is a comprehensive fix, but I did a cursory check of
some obvious places and found these to squash.

+++ src/cineon.imageio/libcineon/Writer.cpp
@@ -304,7 +304,7 @@ bool cineon::Writer::WriteElement(const int element, void *data, const DataSize 
 bool cineon::Writer::WriteThrough(void *data, const U32 width, const U32 height, const int noc, const int bytes, const U32 eolnPad, const U32 eoimPad, char *blank)
 {
 	bool status = true;
-	const int count = width * height * noc;
+	const size_t count = size_t(width) * size_t(height) * noc;
 	unsigned int i;
 	unsigned char *imageBuf = reinterpret_cast<unsigned char*>(data);
